Refactor to make card item reusable for sensors list
