+++ frontend/src/components/Card/card.styles.css
@@ -1,0 +1,44 @@
+.card {
+  background-color: var(--white);
+  margin-bottom: 20px;
+  border-radius: 3px;
+  box-shadow: 2px 4px 12px -1px var(--shadowGrey);
+}
+
+.card__text {
+  color: var(--darkGrey) !important;
+  text-align: left !important;
+}
+
+.card__row {
+  display: flex;
+}
+
+.card__col1 {
+  flex: 1;
+  display: flex;
+  align-items: flex-start;
+  margin-top: 15px;
+  margin-left: 15px;
+}
+
+.card__col2 {
+  flex: 5;
+  margin-left: 15px;
+  margin-top: 15px;
+}
+
+.card__col3 {
+  flex: 1;
+}
+
+.card__row .service-logo {
+  max-width: 100%;
+  max-height: 100%;
+}
+
+@media screen and (max-width: 500px) {
+  .card__col1 {
+    flex: 2;
+  }
+}
